+++ packages/sdk-react-native/src/components/atomic/index.ts
@@ -4,3 +4,4 @@
 
 export { Avatar } from "./Avatar";
 export { AudioIndicator } from "./AudioIndicator";
+export { StatusBadge } from "./StatusBadge";